requirements.txt
openai
python-dotenv
sympy
